Break up command.rs into files per subcommand
