+++ tests/toolchain.rs
@@ -1,0 +1,187 @@
+use anyhow::Result;
+use fuelup::{channel, fmt::format_toolchain_with_target, target_triple::TargetTriple};
+
+mod testcfg;
+use testcfg::{FuelupState, ALL_BINS, DATE};
+
+mod utils;
+use utils::expect_files_exist;
+
+#[test]
+fn fuelup_toolchain_install_latest() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        let output = cfg.fuelup(&["toolchain", "install", "latest"]);
+        assert!(output.status.success());
+
+        for entry in cfg.toolchains_dir().read_dir().expect("Could not read dir") {
+            let toolchain_dir = entry.unwrap();
+            let expected_toolchain_name =
+                "latest-".to_owned() + &TargetTriple::from_host().unwrap().to_string();
+            assert_eq!(
+                expected_toolchain_name,
+                toolchain_dir.file_name().to_str().unwrap()
+            );
+            assert!(toolchain_dir.file_type().unwrap().is_dir());
+        }
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_install_nightly() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        let output = cfg.fuelup(&["toolchain", "install", "nightly"]);
+        assert!(output.status.success());
+
+        for entry in cfg.toolchains_dir().read_dir().expect("Could not read dir") {
+            let toolchain_dir = entry.unwrap();
+            let expected_toolchain_name =
+                "nightly-".to_owned() + &TargetTriple::from_host().unwrap().to_string();
+            assert_eq!(
+                expected_toolchain_name,
+                toolchain_dir.file_name().to_str().unwrap()
+            );
+            assert!(toolchain_dir.file_type().unwrap().is_dir());
+        }
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_install_nightly_date() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        cfg.fuelup(&["toolchain", "install", "nightly-2022-08-31"]);
+
+        for entry in cfg.toolchains_dir().read_dir().expect("Could not read dir") {
+            let toolchain_dir = entry.unwrap();
+            let expected_toolchain_name =
+                "nightly-2022-08-31-".to_owned() + &TargetTriple::from_host().unwrap().to_string();
+            assert_eq!(
+                expected_toolchain_name,
+                toolchain_dir.file_name().to_str().unwrap()
+            );
+            assert!(toolchain_dir.file_type().unwrap().is_dir());
+
+            expect_files_exist(&toolchain_dir.path().join("bin"), ALL_BINS);
+        }
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_install_malformed_date() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        let toolchain = "nightly-2022-08-31-";
+        let output = cfg.fuelup(&["toolchain", "install", toolchain]);
+
+        let expected_stdout = format!("Invalid toolchain metadata within input '{toolchain}' - You specified target '': specifying a target is not supported yet.\n");
+
+        assert!(output.status.success());
+        assert_eq!(output.stdout, expected_stdout);
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_install_date_target_disallowed() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        let toolchain = "nightly-2022-08-31-x86_64-apple-darwin";
+        let output = cfg.fuelup(&["toolchain", "install", toolchain]);
+
+        let expected_stdout =
+            format!("Invalid toolchain metadata within input '{toolchain}' - You specified target 'x86_64-apple-darwin': specifying a target is not supported yet.\n");
+
+        assert!(output.status.success());
+        assert_eq!(output.stdout, expected_stdout);
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_uninstall() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        let toolchains = ["latest", "nightly", &format!("nightly-{}", DATE)];
+        for toolchain in toolchains {
+            let toolchain_with_target = format_toolchain_with_target(toolchain);
+            let output = cfg.fuelup(&["toolchain", "uninstall", toolchain]);
+            let expected_stdout = format!("toolchain '{}' does not exist\n", toolchain_with_target);
+
+            assert_eq!(output.stdout, expected_stdout);
+        }
+    })?;
+
+    testcfg::setup(FuelupState::AllInstalled, &|cfg| {
+        let toolchains = ["latest", "nightly", &format!("nightly-{}", DATE)];
+        for toolchain in toolchains {
+            let toolchain_with_target = format_toolchain_with_target(toolchain);
+            let output = cfg.fuelup(&["toolchain", "uninstall", toolchain]);
+            let expected_stdout = format!("toolchain '{}' uninstalled\n", toolchain_with_target);
+
+            assert!(output.stdout.contains(&expected_stdout));
+        }
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_uninstall_active_switches_default() -> Result<()> {
+    testcfg::setup(FuelupState::LatestAndCustomInstalled, &|cfg| {
+        cfg.fuelup(&["toolchain", "uninstall", "latest"]);
+        let stdout = cfg.fuelup(&["default"]).stdout;
+
+        assert_eq!(stdout, "my-toolchain (default)\n")
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_new() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        let name = "my-toolchain";
+        let output = cfg.fuelup(&["toolchain", "new", name]);
+        let expected_stdout = format!(
+            "New toolchain initialized: {name}
+default toolchain set to '{name}'\n"
+        );
+
+        assert_eq!(output.stdout, expected_stdout);
+        assert!(cfg.toolchain_bin_dir(name).is_dir());
+        let default = cfg.default_toolchain();
+        assert_eq!(default, Some(name.to_string()));
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_new_disallowed() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        for toolchain in [channel::LATEST, channel::NIGHTLY] {
+            let output = cfg.fuelup(&["toolchain", "new", toolchain]);
+            let expected_stderr = format!("error: Invalid value \"{toolchain}\" for '<NAME>': Cannot use distributable toolchain name '{toolchain}' as a custom toolchain name\n\nFor more information try --help\n");
+            assert_eq!(output.stderr, expected_stderr);
+        }
+    })?;
+
+    Ok(())
+}
+
+#[test]
+fn fuelup_toolchain_new_disallowed_with_target() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        let target_triple = TargetTriple::from_host().unwrap();
+        let toolchain_name = "latest-".to_owned() + &target_triple.to_string();
+        let output = cfg.fuelup(&["toolchain", "new", &toolchain_name]);
+        let expected_stderr = format!("error: Invalid value \"{toolchain_name}\" for '<NAME>': Cannot use distributable toolchain name '{toolchain_name}' as a custom toolchain name\n\nFor more information try --help\n");
+        assert_eq!(output.stderr, expected_stderr);
+    })?;
+
+    Ok(())
+}
