+++ tests/version.rs
@@ -1,0 +1,18 @@
+use anyhow::Result;
+use std::env;
+
+mod testcfg;
+use testcfg::FuelupState;
+
+#[test]
+fn fuelup_version() -> Result<()> {
+    testcfg::setup(FuelupState::Empty, &|cfg| {
+        let expected_version = format!("fuelup {}\n", clap::crate_version!());
+
+        let stdout = cfg.fuelup(&["--version"]).stdout;
+
+        assert_eq!(stdout, expected_version);
+    })?;
+
+    Ok(())
+}
